feat(og): tool cards for create-l1 and my-l1
