+++ app/console/create-l1/basic/page.tsx
@@ -1,35 +1,12 @@
-'use client';
+import { consoleToolMetadata } from "@/components/console/tool-metadata";
+import BasicCreateL1Page from "./page.client";
 
-import BasicSetupForm from '@/components/toolbox/console/create-l1/basic/BasicSetupForm';
-import { CheckRequirements } from '@/components/toolbox/components/CheckRequirements';
-import { WalletRequirementsConfigKey } from '@/components/toolbox/hooks/useWalletRequirements';
-import { AccountRequirementsConfigKey } from '@/components/toolbox/hooks/useAccountRequirements';
+// The interactive page is a client component (page.client.tsx), which cannot
+// export metadata; this server wrapper carries the tool's social card. The
+// client file keeps its own 'use client', so its hook-bearing imports
+// (CheckRequirements, wallet requirement hooks) stay out of the server graph.
+export const metadata = consoleToolMetadata("/console/create-l1", "Create L1");
 
-/**
- * Basic Setup entry — gated behind the shared CheckRequirements guard
- * rather than hand-rolling wallet + testnet + login checks inside the
- * form. This matches every other console tool and keeps the messaging,
- * CTAs, and recovery actions consistent across the app.
- *
- * Requirements:
- *   - WalletConnected: owner defaults to the connected wallet, and the
- *     deploy backend needs a real EVM address from the user.
- *   - TestnetRequired: Basic Setup is Fuji-only for MVP; the
- *     orchestrator rejects mainnet requests anyway.
- *   - UserLoggedIn: every deploy is linked to a Builder Hub account
- *     (quick-l1 /deploy rejects requests without a session-derived
- *     userId — see lib/quick-l1/types.ts).
- */
 export default function Page() {
-  return (
-    <CheckRequirements
-      toolRequirements={[
-        WalletRequirementsConfigKey.WalletConnected,
-        WalletRequirementsConfigKey.TestnetRequired,
-        AccountRequirementsConfigKey.UserLoggedIn,
-      ]}
-    >
-      <BasicSetupForm />
-    </CheckRequirements>
-  );
+  return <BasicCreateL1Page />;
 }
